requirements.txt
profanity-check
bjoern
aiohttp
cchardet
brotlipy
python-dotenv
